requirements.txt
streamlit
google-generativeai
Pillow
PyPDF2